requirements.txt
Django==1.7
dj-database-url==0.3.0
django-appconf==0.6
django-compressor==1.4
django-debug-toolbar==1.2.1
django-extensions==1.4.0
psycopg2==2.5.4
six==1.8.0
sqlparse==0.1.12
wsgiref==0.1.2
